chore(@clayui/card): update Card component stories
